added query workload

+++ src/art.cc
@@ -13,21 +13,21 @@ using namespace std;
 
 Node* tree = NULL;
 
-void init(long long *arr, int N) {
+void init(int *arr, int N) {
   for (int i = 0; i < N; i++) {
     insert(arr[i]);
   }
   // art_debug = 1;
 }
 
-void insert(long long value) {
+void insert(int value) {
   uint8_t key[8];
   uint64_t value64 = value;
   loadKey(value64, key);
   insert(tree,&tree,key,0,value64,8);
 }
 
-void erase(long long value) {
+void erase(int value) {
   uint8_t key[8];
   uint64_t value64 = value;
   loadKey(value64, key);
@@ -36,14 +36,14 @@ void erase(long long value) {
   assert(!lookup(tree,key,8,0,8));
 }
 
-long long query(long long value) {
+int query(int value) {
   uint64_t value64 = value;
   uint8_t key[8];
   loadKey(value64, key);
    
   // if (value == 61554031375105761) art_debug = 1;
   Node* leaf=lower_bound(tree,key,8,0,8);
-  long long ret = 0;
+  int ret = 0;
   if (isLeaf(leaf)) {
     ret = getLeafValue(leaf);
     // fprintf(stdout, "%lld (%lld)\n", ret, value);
